Updated Request Service Button.
The button was causing problems (I think) with people not knowing what it did (possibly even that it was even there in the first place). To fix this, I simply changed the hovering button to a simple rectangular one.

TODO: Implement mobile version of the button.

Also cleared up a ton of unused css code.